ict-ai-analyst/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
anthropic==0.40.0
playwright==1.48.0
aiohttp==3.10.0
httpx==0.27.0
Pillow==10.4.0
structlog==24.4.0
